# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NEPZIA Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List/search listings
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: location
          in: query
          schema: { type: string }
        - name: condition
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: sort
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Listings page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsPage"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/featured:
    get:
      operationId: getFeaturedListings
      tags: [listings]
      summary: Get featured/boosted listings
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Featured listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /listings/recent:
    get:
      operationId: getRecentListings
      tags: [listings]
      summary: Get recently posted listings
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Recent listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /listings/nearby:
    get:
      operationId: getNearbyListings
      tags: [listings]
      summary: Get listings near a lat/lng coordinate
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lng
          in: query
          required: true
          schema: { type: number }
        - name: radius
          in: query
          description: Radius in km (default 50)
          schema: { type: number }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Nearby listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /listings/{id}/similar:
    get:
      operationId: getSimilarListings
      tags: [listings]
      summary: Get similar listings
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Similar listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /listings/{id}/report:
    post:
      operationId: reportListing
      tags: [listings]
      summary: Report a listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
...
  /admin/listings:
    get:
      operationId: adminListListings
      tags: [listings]
      summary: Admin — list all listings
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: All listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /admin/listings/{id}/status:
    patch:
      operationId: adminUpdateListingStatus
      tags: [listings]
      summary: Admin — update listing status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingStatusUpdate"
      responses:
        "200":
          description: Updated
components:
  schemas:
...
    Listing:
      type: object
      required: [id, title, price, condition, category, location, status, createdAt, sellerId]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        location: { type: string }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        status: { type: string }
        featured: { type: boolean }
        boosted: { type: boolean }
        viewCount: { type: integer }
        sellerId: { type: string }
        sellerName: { type: ["string", "null"] }
        sellerAvatar: { type: ["string", "null"] }
        sellerVerified: { type: boolean }
        sellerRating: { type: ["number", "null"] }
        contactPhone: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
    ListingsPage:
      type: object
      required: [listings, total, page, limit]
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    ListingInput:
      type: object
      required: [title, price, condition, category, location]
      properties:
        title: { type: string }
        description: { type: string }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        location: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        images: { type: array, items: { type: string } }
        contactPhone: { type: string }
        status: { type: string }
        featured: { type: boolean }
    ListingUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        location: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        images: { type: array, items: { type: string } }
        contactPhone: { type: string }
        status: { type: string }
        featured: { type: boolean }
    ListingStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
    ReportInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
        details: { type: string }
...